Removed conftest.py and moved fixtures to test_bids_utils file
